Restructure JS files
