+++ web_src/js/admin/common.js
@@ -1,7 +1,7 @@
 import $ from 'jquery';
-import {checkAppUrl} from '../common-global.js';
-import {hideElem, showElem, toggleElem} from '../../utils/dom.js';
-import {POST} from '../../modules/fetch.js';
+import {checkAppUrl} from '../features/common-global.js';
+import {hideElem, showElem, toggleElem} from '../utils/dom.js';
+import {POST} from '../modules/fetch.js';
 
 const {appSubUrl} = window.config;
