+++ web_src/js/comp/LabelEdit.js
