+++ web_src/js/comp/QuickSubmit.js
