+++ web_src/js/components/DiffFileList.vue
@@ -1,5 +1,5 @@
 <script>
-import {loadMoreFiles} from '../features/repo-diff.js';
+import {loadMoreFiles} from '../repo/diff.js';
 import {diffTreeStore} from '../modules/stores.js';
 
 export default {
