+++ web_src/js/fomantic/checkbox.js
