+++ web_src/js/repo/diff.js
@@ -1,11 +1,11 @@
 import $ from 'jquery';
-import {initCompReactionSelector} from './comp/ReactionSelector.js';
-import {initRepoIssueContentHistory} from './repo-issue-content.js';
-import {initDiffFileTree} from './repo-diff-filetree.js';
-import {initDiffCommitSelect} from './repo-diff-commitselect.js';
-import {validateTextareaNonEmpty} from './comp/ComboMarkdownEditor.js';
-import {initViewedCheckboxListenerFor, countAndUpdateViewedFiles, initExpandAndCollapseFilesButton} from './pull-view-file.js';
-import {initImageDiff} from './imagediff.js';
+import {initCompReactionSelector} from '../comp/ReactionSelector.js';
+import {initRepoIssueContentHistory} from './issue-content.js';
+import {initDiffFileTree} from './diff-filetree.js';
+import {initDiffCommitSelect} from './diff-commitselect.js';
+import {validateTextareaNonEmpty} from '../comp/ComboMarkdownEditor.js';
+import {initViewedCheckboxListenerFor, countAndUpdateViewedFiles, initExpandAndCollapseFilesButton} from '../features/pull-view-file.js';
+import {initImageDiff} from '../features/imagediff.js';
 import {showErrorToast} from '../modules/toast.js';
 import {submitEventSubmitter, queryElemSiblings, hideElem, showElem} from '../utils/dom.js';
 import {POST, GET} from '../modules/fetch.js';
